requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==41.0.7
reportlab==4.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
